refactor code and move components in Today to separate files
